fix merge conflict

+++ app/Http/Controllers/Admins/GroupController.php
@@ -9,6 +9,12 @@ use App\Models\Survey;
 
 class GroupController extends Controller
 {
+    public function __construct()
+    {
+        // to middleware or not to middleware?
+        // $this->middleware('group');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -30,6 +36,8 @@ class GroupController extends Controller
      */
     public function create(Survey $survey)
     {
+        $this->authorize('resource', $group);
+
         return view('admins.groups.create')
             ->with('survey', $survey)
             ->with('group', new Group);
@@ -43,6 +51,8 @@ class GroupController extends Controller
      */
     public function store(GroupForm $request, Survey $survey)
     {
+        $this->authorize('resource', $group);
+
         $survey->groups()->create([
             'slug' => $request->slug,
             'title' => $request->title,
@@ -61,7 +71,9 @@ class GroupController extends Controller
      */
     public function show(Survey $survey, Group $group)
     {
-        $group = $survey->groups()->where('groups.id', $group->id)->first();
+        $this->authorize('resource', $group);
+
+        $group = $survey->groups()->where('groups.id', $group->id)->firstOrFail();
         
         return view('admins.groups.show')
             ->with('survey', $survey)
@@ -76,6 +88,8 @@ class GroupController extends Controller
      */
     public function edit(Survey $survey, Group $group)
     {
+        $this->authorize('resource', $group);
+
         return view('admins.groups.edit')
             ->with('survey', $survey)
             ->with('group', $group);
@@ -90,6 +104,8 @@ class GroupController extends Controller
      */
     public function update(GroupForm $request, Survey $survey, Group $group)
     {
+        $this->authorize('resource', $group);
+
         $group->update([
             'slug' => $request->slug,
             'title' => $request->title,
@@ -108,6 +124,8 @@ class GroupController extends Controller
      */
     public function destroy(Survey $survey, Group $group)
     {
+        $this->authorize('resource', $group);
+
         $group->delete();
 
         return redirect()->route('surveys.groups.index', [$survey->id]);
