+++ app/Http/Controllers/Admins/SurveyController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers\Admins;
 
+use Illuminate\Http\Request;
+
 use App\Http\Controllers\Controller;
 use App\Http\Requests\SurveyForm;
 use App\Models\Survey;
@@ -14,9 +16,10 @@ class SurveyController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        $surveys = Survey::all();
+        $surveys = $request->user()->surveys()->get();
+
         return view('admins.surveys.index')
             ->with('surveys', $surveys);
     }
@@ -28,6 +31,8 @@ class SurveyController extends Controller
      */
     public function create()
     {
+        $this->authorize('resource', $survey);
+
         return view('admins.surveys.create')
             ->with('survey', new Survey);
     }
@@ -40,6 +45,8 @@ class SurveyController extends Controller
      */
     public function store(SurveyForm $request)
     {
+        $this->authorize('resource', $survey);
+
         Survey::create([
             'slug' => $request->slug,
             'title' => $request->title,
@@ -66,6 +73,8 @@ class SurveyController extends Controller
      */
     public function show(Survey $survey)
     {
+        $this->authorize('resource', $survey);
+
         return view('admins.surveys.show')
             ->with('groups', $survey->groups)
             ->with('survey', $survey);
@@ -79,6 +88,8 @@ class SurveyController extends Controller
      */
     public function edit(Survey $survey)
     {
+        $this->authorize('resource', $survey);
+
         return view('admins.surveys.edit')
             ->with('survey', $survey);
     }
@@ -92,6 +103,8 @@ class SurveyController extends Controller
      */
     public function update(SurveyForm $request, Survey $survey)
     {
+        $this->authorize('resource', $survey);
+
         $survey->update([
             'slug' => $request->slug,
             'title' => $request->title,
@@ -118,6 +131,8 @@ class SurveyController extends Controller
      */
     public function destroy(Survey $survey)
     {
+        $this->authorize('resource', $survey);
+
         $survey->delete();
 
         return redirect()->route('surveys.index');
